requirements.txt
pydantic==2.11.5
pydantic-settings==2.9.1
fastapi==0.115.9
langgraph==0.4.1
langchain==0.3.25
langchain-core==1.2.16
langchain-openai==1.1.10
langchain-community==0.3.24
openai==2.24.0
python-dotenv==1.0.1
psycopg==3.1.19
psycopg2-binary==2.9.10
SQLAlchemy==2.0.40
alembic==1.14.1
pgvector==0.3.6
pymupdf
